// packages/fe-job-pc/src/pages/admin/service/hierarchical-diagnosis/diagnosis/interface.d.ts
export interface AssistanceFromQuestionDto {
    /**
     * 题目编码
     */
    code?: string;
    /**
     * 引导语
     */
    guide?: string;
    /**
     * 模块编码
     */
    moduleCode?: string;
    /**
     * 模块标题
     */
    moduleTitle?: string;
    /**
     * 选项列表
     */
    options?: AssistanceFromQuestionOptionDto[];
    /**
     * 段落编码
     */
    paragraphCode?: string;
    /**
     * 段落标题
     */
    paragraphTitle?: string;
    /**
     * 题干
     */
    title?: string;
    /**
     * 措辞优化
     */
    wordingOptimization?: string;
    [property: string]: any;
}

/**
 * 选项
 *
 * AssistanceFromQuestionOptionDto
 */
export interface AssistanceFromQuestionOptionDto {
    /**
     * 选项描述
     */
    description?: string;
    /**
     * 选项等级 1弱 2一般 3强
     */
    level?: number;
    /**
     * 选项分值
     */
    score?: number;
    /**
     * 选项编码
     */
    sequence?: string;
    /**
     * 选项内容
     */
    title?: string;
    [property: string]: any;
}